week_1/requirement.txt
vosk
openai-whisper
jiwer
soundfile
torch
